LayoutTests/js/dom/prototype-chain-caching-with-impure-get-own-property-slot-traps-llint-expected.txt

Tests what happens when you make prototype chain accesses with impure GetOwnPropertySlot traps in the way.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "function"
PASS typeof result is "object"
PASS typeof result is "object"
PASS typeof result is "object"
PASS typeof result is "object"
PASS typeof result is "object"
PASS successfullyParsed is true

TEST COMPLETE
